feat: remove new db

+++ src/core/cache.lua
@@ -5,15 +5,9 @@ local log = require "src.utils.log"
 local config = require "src.core.config"
 
 local M = {}
+local limit = 3600
 
-function M:new()
-  return setmetatable({
-    limit = 3600,
-    db = {},
-  }, { __index = self })
-end
-
-function M:save(name, cache)
+function M.save(name, cache)
   local C = {
     timestamp = os.time(),
     cache = cache,
@@ -21,50 +15,32 @@ function M:save(name, cache)
 
   local text = "return " .. fn.inspect(C)
   local file = fs.join(PKGER_CACHE, name)
-
-  self.db[name] = cache
-
   return fs.write_file(file, text)
 end
 
-function M:load(name, flags)
-  if self.db[name] and (os.time() - (self.db[name].timestamp or 0) < self.limit) then
-    return self.db[name].cache
-  end
-
+function M.load(name, flags)
   local file = fs.join(PKGER_CACHE, name)
   local ok, C = pcall(dofile, file)
 
-  if ok and (os.time() - (C.timestamp or 0) < self.limit) then
-    self.db[name] = C
+  if ok and (os.time() - (C.timestamp or 0) < limit) then
     return C.cache
   end
 
   return nil
 end
 
-function M:clear(name)
+function M.clear(name)
   local file = fs.join(PKGER_CACHE, name or "")
 
-  if name then
-    self.db[name] = nil
-    if fs.is_file(file) then
-      return fs.rm(file)
-    end
+  if name and fs.is_file(file) then
+    return fs.rm(file)
   end
 
-  if not name then
-    self.db = {}
-    if fs.is_dir(PKGER_CACHE) then
-      return fs.rm_dir(PKGER_CACHE)
-    end
+  if not name and fs.is_dir(PKGER_CACHE) then
+    return fs.rm_dir(PKGER_CACHE)
   end
 
   return false, nil
 end
-
--- function M:print()
---   print(self.limit)
--- end
 
 return M
